refactor: Move IngesterQueryRequest to data_types2
So that querier doesn't need to depend on ingester.
