+++ generated_types/Cargo.toml
@@ -7,9 +7,12 @@ edition = "2021"
 [dependencies] # In alphabetical order
 bytes = "1.0"
 data_types = { path = "../data_types", optional = true }
+data_types2 = { path = "../data_types2", optional = true }
+datafusion = { path = "../datafusion", optional = true }
 observability_deps = { path = "../observability_deps" }
 pbjson = "0.2"
 pbjson-types = "0.2"
+predicate = { path = "../predicate", optional = true }
 prost = "0.9"
 regex = "1"
 serde = { version = "1.0", features = ["derive"] }
@@ -28,4 +31,4 @@ pbjson-build = "0.2"
 
 [features]
 default = ["data_types_conversions"]
-data_types_conversions = ["data_types"]
+data_types_conversions = ["data_types", "data_types2", "datafusion", "predicate"]
