+++ generated_types/src/lib.rs
@@ -214,6 +214,8 @@ pub mod database_state;
 #[cfg(any(feature = "data_types_conversions", test))]
 pub mod delete_predicate;
 #[cfg(any(feature = "data_types_conversions", test))]
+pub mod ingester;
+#[cfg(any(feature = "data_types_conversions", test))]
 pub mod job;
 #[cfg(any(feature = "data_types_conversions", test))]
 pub mod router;
