+++ predicate/src/lib.rs
@@ -10,7 +10,6 @@
 
 pub mod delete_expr;
 pub mod delete_predicate;
-pub mod ingester;
 pub mod regex;
 pub mod rewrite;
 pub mod rpc_predicate;
